pyproject.toml
[project]
name = "modbus-simulator"
version = "1.0.0"
description = "A powerful Modbus device simulator with web monitoring interface"
authors = [
    {name = "Your Name", email = "your.email@example.com"}
]

dependencies = [
    "pymodbus>=3.5.2",
    "fastapi>=0.110.0",
    "uvicorn[standard]>=0.27.1",
    "websockets>=12.0",
    "jinja2>=3.1.3",
    "python-multipart>=0.0.9",
    "python-socketio>=5.11.1",
    "aiohttp>=3.9.3",
    "rich>=13.7.1",
    "orjson>=3.9.15",
    "httptools>=0.6.1",
    "watchfiles>=0.21.0",
    "python-jose[cryptography]>=3.3.0",
    "passlib[bcrypt]>=1.7.4",
    "python-dotenv>=1.0.1",
]
requires-python = ">=3.11"
readme = "README.md"
license = {text = "MIT"}

[tool.hatch.metadata]
allow-direct-references = true

[project.scripts]
modbus-simulator = "src.main:main"

[tool.ruff]
line-length = 120
target-version = "py311"

[tool.ruff.lint]
select = ["E", "F", "B", "I"]
ignore = []

[tool.ruff.format]
quote-style = "double"
indent-style = "space"
line-ending = "auto"
